Add support for hostnames without punctuation
This adds support for loading modules on a local network or via IP
addresses.

[extract_repo_parts]

* Change  domain part general check for considering punctuation on
domain name parts. When that is the case, add a check for requiring
mandatory extension to the name.

* Change "domain provided" code path to always expect for explicit
protocol. Also add the punctuation check as well.

* Change "implicit protocol" code path regex for splitting username and
reponame to be more specific with regards to the hostname part instead
of (.*)

* Change some error messages for better understanding